% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bSCMdeConfoundingGraphFunc.R
\name{bSCMdeConfoundingGraphFunc}
\alias{bSCMdeConfoundingGraphFunc}
\title{bSCMdeConfoundingGraphFunc function}
\usage{
bSCMdeConfoundingGraphFunc(dat, IndpThs = 0.05, alpha = 0.05)
}
\description{
bSCMdeConfoundingGraphFunc function
}
