% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bSCMDepndentGraphFastFunc.R
\name{bSCMDepndentGraphFastFunc}
\alias{bSCMDepndentGraphFastFunc}
\title{bSCMDepndentGraphFunc function}
\usage{
bSCMDepndentGraphFastFunc(mat, IndpThs = 0.05)
}
\description{
bSCMDepndentGraphFunc function
}
