% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bSCMDepndentGraphFunc.R
\name{bSCMDepndentGraphFunc}
\alias{bSCMDepndentGraphFunc}
\title{bSCMDepndentGraphFunc function}
\usage{
bSCMDepndentGraphFunc(mat, nboot = 100, alpha = 0.05, IndpThs = 0.05)
}
\description{
bSCMDepndentGraphFunc function
}
