% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causalDirTest.R
\name{causalDirTest}
\alias{causalDirTest}
\title{causalDirTest function}
\usage{
causalDirTest(
  mat,
  i,
  j,
  vi = 1,
  vj = 1,
  z = c(),
  alpha = 0.05,
  ths = 0.25,
  nboot = 100,
  slack = 0.001
)
}
\description{
causalDirTest function
}
