% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparePredAdjMatrix2TrueAdjMat.R
\name{comparePredAdjMatrix2TrueAdjMat}
\alias{comparePredAdjMatrix2TrueAdjMat}
\title{comparePredAdjMatrix2TrueAdjMat}
\usage{
comparePredAdjMatrix2TrueAdjMat(trueAdjMat, adjMat)
}
\arguments{
\item{trueAdjMat}{a ground-truth matrix.}

\item{adjMat}{an inferred matrix.}
}
\value{
This function returns a list of precision \code{prec}, recall \code{rec}, and F1 score \code{F1} of inferred vs. groundtruth matrices.
}
\description{
comparePredAdjMatrix2TrueAdjMat is a support function that can compare two adjacency matrices: ground-truth and inferred matrices.
}
\examples{
## Generate simulation data
#G<-matrix(FALSE,10,10) # groundtruth
#G[1,c(4,7,8,10)]<-TRUE
#G[2,c(5,7,9,10)]<-TRUE
#G[3,c(6,8,9,10)]<-TRUE
#checkMultipleSimulationVLtimeseries(trueAdjMat=G,adjMat=G)

}
