% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CondProb.R
\name{CondProb}
\alias{CondProb}
\title{CondProb function
The function serves as a confidence function in data mining conf(y|z)
Given a set of transaction D}
\usage{
CondProb(D, y, z)
}
\description{
CondProb function
The function serves as a confidence function in data mining conf(y|z)
Given a set of transaction D
}
