% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confNetFunc.R
\name{confNetFunc}
\alias{confNetFunc}
\title{confNetFunc function
Computing a confidence network in data mining
where  there is a weight of edge from i to j as conf(j|i)}
\usage{
confNetFunc(mat, ths = 0.1)
}
\description{
confNetFunc function
Computing a confidence network in data mining
where  there is a weight of edge from i to j as conf(j|i)
}
