% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{D}
\alias{D}
\title{An example of aligned list of transactions}
\format{
An aligned list of a matrix with 200 samples and 10 dimensions generated from Bernoulli distribution.
\describe{
\item{D}{It is an aligned list of transactions that was converted from \code{mat}.}
}
}
\usage{
D
}
\description{
A dataset containing simulated data that is used for examples in the package.

The \code{D} is an aligned list of transactions that was converted by using \code{D<-VecAlignment(simData$mat)}.
}
\keyword{datasets}
