% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getReachableNodes.R
\name{getTransitiveClosureMat}
\alias{getTransitiveClosureMat}
\title{getTransitiveClosureMat function}
\usage{
getTransitiveClosureMat(adjMat)
}
\arguments{
\item{adjMat}{is an adjacency matrix of a  directed graph of which its elements are binary: zero for no edge, and one for having an edge.}
}
\value{
This function returns a transitive-closure adjacency matrix.
}
\description{
getTransitiveClosureMat is a support function for inferring a transitive-closure adjacency matrix.
}
\examples{
# Given an example of adjacency matrix
A<-matrix(FALSE,5,5)
A[2,1]<-TRUE
A[c(3,4),2]<-TRUE
A[5,3]<-TRUE
# Get a set of reachable nodes of targetNode.

trsClosureMat<-getTransitiveClosureMat(adjMat=A)

}
