% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustmentProb.R
\name{num2Bits}
\alias{num2Bits}
\title{num2Bits function}
\usage{
num2Bits(num, n = 32)
}
\arguments{
\item{num}{is a natural number.}

\item{n}{is a number of bits representing \code{num}.}
}
\value{
This function returns an n-dimensional vector of bits that represents \code{num}.
}
\description{
Given a natural number and number of bits, the function provides an n-dimensional vector of bits that represents \code{num}.
The ith bits of binary vector represents the ith bit of \code{num}.
For example, if \code{vec<-num2Bits(num=2,n=4)}, the first bit \code{vec[1]} is 0 and the second bit \code{vec[2]} is 1.
}
\examples{
num2Bits(num=10,n=4)

}
