% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{resC}
\alias{resC}
\title{An example of causal inference result}
\format{
A result of causal inference using \code{mat} as an input.
\describe{
\item{resC}{It is a result of causal inference using simData$mat as an input by running
\code{resC<-BiCausality::CausalGraphInferMainFunc(mat = mat,CausalThs=0.1, nboot =50, IndpThs=0.05)}. }.
}
}
\usage{
resC
}
\description{
A dataset containing a result of causal inference from simulated data that is used for examples in the package.
}
\keyword{datasets}
