% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simData}
\alias{simData}
\title{A simulation}
\format{
...
\describe{
\item{simData$mat}{ It is a set of time series where \code{TS[i,t,d]} is a numeric value of \code{i}th time series at time \code{t} and dimension \code{d}.}
...
}
}
\usage{
simData
}
\description{
A dataset containing simulated data...
}
\keyword{datasets}
