% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VecAlignment.R
\name{VecAlignment}
\alias{VecAlignment}
\title{VecAlignment function}
\usage{
VecAlignment(mat)
}
\arguments{
\item{mat}{is a matrix n by d where n is a number of transactions or samples and d is a number of dimensions.}
}
\value{
This function returns  an aligned list of transactions \code{D}, is an aligned list of transactions that was converted from any matrix n by d \code{mat}.
}
\description{
This function rearranges the samples in the \code{mat} into
an aligned list of transactions, which is mainly used by other functions in the package.
Suppose mat\link{i,} is a binary vector  we are interested, we use  \code{A<-bin2dec(mat[i,])}
to store the decimal value of \code{mat[i,]} in \code{A}. Then, we call \code{D[[A]]$count}
to get number of samples in \code{mat} that are similar to \code{mat[i,]} and
the \code{D[[A]]$name} is \code{mat[i,]}.
}
\examples{
VecAlignment(mat=simData$mat)

}
