% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assocSignTest.R
\name{assocSignTest}
\alias{assocSignTest}
\title{indpFunc function}
\usage{
assocSignTest(
  mat,
  i,
  j,
  z = c(),
  alpha = 0.05,
  ths = 0.05,
  nboot = 100,
  slack = 0.001
)
}
\description{
indpFunc function
}
