% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VecAlignment.R
\name{bin2dec}
\alias{bin2dec}
\title{bin2dec function}
\usage{
bin2dec(X)
}
\arguments{
\item{X}{is a binary vector where \code{X[i]} is the ith bit of vector.}
}
\value{
This function returns a decimal value of \code{X}.
}
\description{
This function convertes a binary vector into its decimal value.
}
\examples{
bin2dec(X=c(1,1,1,0))

}
