% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bSCMCausalGraphFunc.R
\name{bSCMCausalGraphFunc}
\alias{bSCMCausalGraphFunc}
\title{bSCMCausalGraphFunc function}
\usage{
bSCMCausalGraphFunc(E1, Dboot, alpha = 0.05, SignThs = 0.05, CausalThs = 0.25)
}
\description{
bSCMCausalGraphFunc function
}
