% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CausalGraphInferMainFunc.R
\name{CausalGraphInferMainFunc}
\alias{CausalGraphInferMainFunc}
\title{CausalGraphInferMainFunc function}
\usage{
CausalGraphInferMainFunc(
  mat,
  alpha = 0.05,
  nboot = 100,
  IndpThs = 0.05,
  CausalThs = 0.1,
  slack = 0.001
)
}
\description{
CausalGraphInferMainFunc function
}
